Do not release the PinotDataBuffer when closing the index (#5400)
The PinotDataBuffers are tracked and maintained inside SegmentDirectory
for ImmutableSegment and PinotDataBufferMemoryManager for MutableSegment.
They are created when loading the indexes and released when closing the
segment. If the PinotDataBuffer gets released when closing the indexes,
and if the buffer manager decides to reuse the buffer, the following read
on the buffer will cause JVM to crash. This can be triggered in
SegmentPreProcessor when the same indexes need to be opened twice in two
different preprocessors.

This PR standardize the behavior of indexes to not release (close) the
PinotDataBuffer when closing the index.
We don't need to worry about accessing the index after it is already
closed because that is already addressed in #4764