Fix issue with PerfBenchmarkRunner where server can no longer pre-load segments from dataDir.
Due to recent PRs (#5383, #5381), PerfBenchmarkRunner now makes additional copy
of segments on controller, and makes server download data from controller.
The original behavior was to bypass this and for servers to directly preload
from disk. This PR is undoing the changes from above PR to get back to old
behavior that is more optimal.