Refactor DistinctTable to use PriorityQueue based algorithm
Currently DISTINCT query is solved the same way as GROUP-BY queries,
which is not necessary (consume much more memory and CPU) and does
not guarantee accuracy of the result.

Instead, DISTINCT query can be solved by a set and a heap efficiently
(similar to SelectionOrderBy but unique records need to be tracked).

The new DistinctTable does not implement the Table interface because
the table interface is designed mostly for the GROUP-BY queries, and
is not efficient for DISTINCT. If in the future we want to use Table
interface to uniform all the input/output, we can redesign the Table
interface to make it suitable for all types of queries.