Changed the segment commit protocol to send/receive streamPartitionMsgOffset
Updated the segment commit protocol so that new element streamPartitionMsgOffset
is populated in requests (as request parameters) and in response (as JSON string element)

The server side has been modified to send the 'streamPartitionMsgOffset' as well as we the
'offset' parameters to the controller. The controller looks for and prefers streamPartitionMsgOffset
but falls back to offset if the streamPartitionMsgOffset is not there.

The controller side, in the repsonse message, populates both of the elements, and the server on
the receiver side does likewise -- preferring streamPartitionMsgOffset.

All callers into the protocol module have been modified to NOT set the offset field. Instead,
only set the streamPartitionMsgOffset field. The 'offset' value will be derived from
streamPartitionMsgOffset.

Added a test to make sure that the controller always generates both elements. Such a test was not
possible in the server side at this time, so verified manually.

Manually ran LLCClusterIntergrationTest by disabling populating `streamPartitionMsgOffset` on the
server side (old server/new controller) and on the controller respons side (new server/old controller)

Issue #5359