Fix travis_quickstart.sh to handle delayed cluster setup. (#5498)
In case of delay in cluster setup, the script fails with syntax error
because the variable that stores the query result is either null or '',
which ends up being compared to a number.

The fix here is to add a check ensure that the query result is indeed
a number, before comparing it against a number. In case of delayed start,
the existing loop of trying 200 times (with a one second sleep) should
already suffice.