Remove BlockValIterator and directly read values from BlockValSet with docId
When we read values from BlockValIterator, we don't use it as an iterator.
Instead, we usually call skipTo(docId) followed by the actual read. This
has the overhead of one extra method invocation, and is also counter
intuitive.
This PR enhanced the BlockValSet to provide APIs to directly read value
for a given docId. This can provide better performance, and better
abstraction of value reads.
Note that the BlockValSet here is not actually for a block, but for a
whole column. The following PRs will try to split the BlockValSet into
multiple interfaces to differentiate block level value set and segment
level value set.