Fix the failure caused by Reflections in FunctionRegistry
The problem is caused by Reflections library not being thread-safe when multiple threads are accessing the same jar file.
Read more about the thread-safety issue here: https://github.com/ronmamo/reflections/issues/81
Users are reporting the same issue even with the latest Reflections version 0.9.12.
A common solution is to downgrade Reflections to 0.9.9, but we have other dependencies rely on 0.9.11, so downgrading might cause other issues.

The solution introduced here is by replacing the static block with an init() method so that we can control when to scan the files to avoid the thread-safety issue.
Another benefit of using an init() method is that the exception won't be swallowed (the exception in static block will be swallowed and query engine will start getting ClassNotFoundException)

This PR also introduces a convention for the plugin methods using reflection, where the class must includes ".function." in its class path.
This can significantly reduce the time of class scanning (reduced from 4 seconds to 200 ms locally)