Fix the failure caused by Reflections in FunctionRegistry (#5531)
The new FunctionRegistry change of using reflection to plug in methods (introduced in #5440) is causing failure in query compilation.
The problem is caused by Reflections library not being thread-safe when multiple threads are accessing the same jar file.
We have multiple libraries (jersey, swagger) using reflection to load classes. The FunctionRegistry uses reflection in the static block, which happens when the class is loaded. If the first query arrives (first usage of FunctionRegistry which runs the static block) during the setup of the jersey server, Reflections will throw exception.
Using reflection in the static block can also cause the first query to block on reflection scanning the classes, which can potentially take seconds.
Read more about the thread-safety issue here: ronmamo/reflections#81
Users are reporting the same issue even with the latest Reflections version 0.9.12.
A common solution is to downgrade Reflections to 0.9.9, but we have other dependencies rely on 0.9.11, so downgrading might cause other issues.

The solution here is to introduce a no-op init() method to trigger the static block so that we can control when to scan the files to avoid the thread-safety issue as well as blocking the first query.
Another benefit of using an init() method is that the exception won't be swallowed (the exception in static block will be swallowed and query engine will start getting ClassNotFoundException)

This PR also introduces a convention for the plugin methods using reflection, where the class must includes ".function." in its class path.
This can significantly reduce the time of class scanning (reduced from 4 seconds to 200 ms locally)