Change default segment load mode to MMAP.
The load mode for segments currently defaults to `heap`. This PR
changes the load-mode to default to `mmap` instead. This is for
two reasons:

1. We have been primarly using and recommending `mmap` mode, so no
   reason why the default should be `heap`.
2. Heap mode can lead to OOMs.

Also added a log message upon successful loading of segments along
with the load mode. This was because, in absence of the log message,
we see the following misleading message:

```
        Instance Data Dir: ....
        Instance Segment Tar Dir: ....
        Bootstrap Segment Dir: null
        Read Mode: heap
```