Remove '/docker', '/kubernetes', '/website' for src release
According to #5540, our source release included some binary files
and it has been flagged during the review process for the release.

We have decided to exclude some dirs from the source release because
docker or kubernetes scripts are more for helping users to easily
try out the Pinot and not the essential part of the source release.