Handle case sensitivity for PinotQuery within BrokerRequest
Currently the case sensitivity handling only applies to the BrokerRequest, but not to the PinotQuery inside the BrokerRequest.
This will cause inconsistent and unexpected behavior when the query engine uses PinotQuery.

Fix this issue by extending the case sensitivity handling to the PinotQuery if it exists.
Note that because SQL queries are also compiled to BrokerRequest, the case sensitivity handling already applies to both PQL and SQL.

Deprecate the config key `enable.case.insensitive.pql` in Helix cluster config and replace it with `enable.case.insensitive`, but keep backward compatibility.