Add QueryContext to replace BrokerRequest in the query engine (#5483)
Motivation:
For historic reason, BrokerRequest does not support expressions natively.
In order to support expressions (transform), the current solution is to
store the expressions as String and compile them when constructing the
operators.
The problem with this approach is that the same compilation is performed
multiple times on EACH segment. This could cause performance degradation
when each server is hosting lots of segments.

In this PR, we introduced a new Object - QueryContext to encapsulate all
the query related information to be used by the query engine. It is
already fully compiled and can be shared by all the segments.
The reason of introducing the new QueryContext over using the PinotQuery
is that with QueryContext:
- It is not used in the wiring layer so that execution layer can be
  decoupled from the wiring layer, and changes for one layer won't affect
  the other layer.
- It is very hard to change wiring Object (PinotQuery) because it involves
  protocol change, so we should make it as generic as possible to support
  future features. Instead, QueryContext can be upgraded along with the
  new future support in query engine as needed, and we don't have to make
  it very generic, which can help save the overhead of handling generic
  Objects.
- In case we need to change the wiring Object (e.g. switch from Thrift to
  Protobuf), we don't need to change the whole query engine.
- We can add some helper variables or methods in the context classes which
  can be shared for all segments to reduce the repetitive work for each
  segment.

This PR only introduces the new QueryContext, the related context classes
and a Util class to get QueryContext from BrokerRequest. The following PRs
will make incremental changes to move the query engine to use the new
QueryContext.