Re-enable S3PinotFSTest.
The test was disabled because it was causing failures in our testing environment
as S3MockListener was starting S3Mock at a port that was already in use, and there
was no way to reconfigure that port. Now that we are no longer running tests in that
environment, re-enabling the test.