Add error message to broker response in case of broker send error (#5705)
On Pinot Broker, the incoming request gets written into socket channels of the target Servers. This happens on `QueryRouter.submitQuery(...)` function. If any exception occurs during  submitQuery for any reason like connection refused to one of the servers, sending requests to the remaining servers are abandoned and the partial responses from already successful sent requests are returned in BrokerResponse with no indication of the exception.
Although partial response is acceptable, there should be an indication of such problem in broker response to make life easier for ppl debugging this issue. Recently there was an incident where, for a specific use case, broker response was empty (no partial response) and there was no exception returned, while data was available on Pinot Servers. After good amount of time going through logs, this issue was discovered with exception being connection refused by only one faulty server.
This PR adds the exception to BrokerReponse for easier debugging.