Allow serving of offline segments immediately/inclusively.
Today, Pinot serves queries to Hybrid tables by applying a time filter
based on data availability. For example, when the most recent offline
segment is for 1/2/2000, and a query has no time filter, then the broker
queries the offline table for * - 1/2/2000 (exclusive), and the realtime
table with 1/2/2000 (inclusive) - *.

This means that when you upload a new offline segment, it technically
doesn't begin to be served by the broker until another offline segment
pushes the watermark higher.

This change permits a new config which tells the broker to immediately
start serving offline segments once available; or in other words, to
query the offline table with the most recent time value inclusively.