Avoid redundant merge of empty sketches.
For segments where no rows are selected, the intermediate and the final merge steps
still create empty Union objects and merge them. This can cause huge latency degradation
in cases there is a large number of segments, and large value of nominal entriesf

This PR avoids creation/merging of empty sketches. Doing so, we are seeing a latency
improvement from 8s to < 500ms for 167 segments, and nominal entries of `1048576`.