Lazy initialization of ForwardIndexReaderContext.
Reader context is currently getting eager-initialized during operator initialization.
In certain cases this leads to OOM for direct memory as follows:

- Consider a case where we have hundreds of segments, but the rows of interest typically
  are within a much smaller subset of segments.

- For cases with large bytes blobs (theta-sketches), the ForwardIndexReaderContext can initialize
  chunks (from direct-memory) of large size (numRowsPerChunk * lengthOfLargestEntry). This can run into
  several tens of MBs for theta-sketches.

- Under such case, even though the rows of interest are in a small set of segments, the ForwardIndexReader
  initializes huge chunks for hundreds of segments, which is wasteful.

This PR fixes the problem with lazy initialization of the ForwardIndexReaderContext, so that it only
gets initialized for segments that have the data.

Existing tests are expected to provide coverage for this code change.