[TE] The endpoint for searching anomalies and pagination (#5778)
This PR creates the endpoints for searching and paginate the anomalies. The endpoint is built for the anomaly page v3. It brings the response time from ~30s to around 300ms. We should see a ~100x speedup of the page load time.

The current Anomaly list page is loading slowly because of calling legacy endpoints. Although the old endpoint works ok when the scale is small, it responses very slowly when handling anomalies on a large scale. (above 100k+ anomalies)

This endpoint pushes the filtering and pagination of the anomalies into backend and database servers instead of doing it in the UI.