update Swagger (OpenAPI) configuration for HTTP+HTTPS (#5817)
Before, the Swagger initialization was only permitting HTTP schema for
most applications and optionally the HTTPS schema which had to be
configured.

Now, Swagger will always list both schemas. By default, it picks the
first schema (which will be HTTP) on the client. When a Swagger page is
opened the swagger JS client will prefer the protcol the page is on, if
it is listed as supported. This means the Swagger browser pages for
Pinot will now work out-of-the-box for HTTPS-only setups.

For this to work though, Swagger UI needs to be updated because the
current version does not correctly select the scheme, which was addessed
in https://github.com/swagger-api/swagger-js/pull/857.