[TE] add anomaly detection as a service - Phase 1 (#5769)
This PR creates three endpoints for providing anomaly detection as the REST API service. Endpoints are provided in a new class called AnomalyDetectionResource.

The 1st endpoint is /anomaly-detection which allows users to send requests with ad-hoc data (JSON file), run AD task synchronously, and return anomalies after task completed. A new class called OnlineThirdEyeDataSource is created.
The 2nd endpoint is /anomaly-detection/task which allows users to send requests to run an AD task via existing configurations asynchronously and will return a task ID.
The 3rd endpoint is /anomaly-detection/{taskId} which allows users querying task status.
This is phase 1 for this feature. During phase 2, a new table for ad-hoc data will be created. And logic will be optimized to have shorter querying time for the ad-hoc endpoint.

Testings are added for some methods in AnomalyDetectionResource. More testings will be added in phase 2.