Add RANGE FilterKind to support merging ranges for SQL (#5898)
Currently we cannot replace `BrokerRequest` with `PinotQuery` completely because of the filter optimizer only works on `BrokerRequest`. One of the optimizations is to merge ranges joined by AND (very useful especially for hybrid table).
We cannot apply the same optimization to `PinotQuery` without a new `RANGE` FilterKind (e.g. no way to represent a > 10 AND a <= 20).
This PR introduces the new `RANGE` FilterKind so that Server knows how to process it.
After release `0.5.0`, we can add the optimizer for `PinotQuery` so that Broker can optimize `PinotQuery`, and Server already knows how to process the `RANGE` filter.

Also let Server support both `'\0'` and `"\t\t"` as the range delimiter so that it can work on string values with `'\t'` inside properly.