[TE] add anomaly detection as a service - endpoints for managing online detection data (#5894)
Phase 2 is separated into two parts. This PR is for the 2nd part. Please note that this PR has a dependency. PR #5866 should be merged before this PR get merged. The PR for Phase 2 1st part can be found in #5866 (this is also the commit c118b16 in current PR). The PR for Phase 1 can be found in #5769.

The main change is that several CRUD endpoints are created for managing online detection data. Those endpoints are listed as follows:

* POST /anomaly-detection/data - register data. The data should be provided in the request. The dataset and metric configurations can optionally be provided to customize configurations. A data-id will be returned.
* GET /anomaly-detection/data/{data-id} - retrieve registered data with the corresponding dataset and metric configurations.
* PUT /anomaly-detection/data/{data-id} - update registered data. The data should be provided in the request. The dataset and metric configurations can optionally be provided to customize configurations. The data-id will be returned.
* DELETE /anomaly-detection/data/{data-id} - delete registered data. The data-id will be returned.

And one extra optional path parameter data-id is provided for online detection service endpoint: POST /anomaly-detection. If the user has used the above endpoints to register their online detection data, they do not need to provide data in the online detection service request (POST /anomaly-detection) and only need to provide the data-id with optional detection configuration.

Here are some other minor changes:
* Add dataset and metric configurations in response as well (similar to detection configuration)
* Force user to provide start and end time. If they are not provided, return with error message
* The parameters used in request/response are all reformated to use hyphen format. For example, detectionConfiguration is changed to detection-configuration.