[TE] add anomaly detection as a service - new table for storing ad-hoc data (#5866)
Phase 2 is separated into two parts. This PR is for the 1st part. The main change is that a new table is created for storing online ad-hoc data. The PR for Phase 1 can be found in #5769.

Here are some other minor changes:
* Remove the cleanExistingState method in the AnomalyDetectionResource. This is not necessary because each request is distinct now and the monitor job will periodically clean them (dataset/metric/detection configs, ad-hoc data).
* Remove the suffix appended to metric as the metric name is allowed to be duplicate.

The 2nd part of phase 2 is about creating a new endpoint to provide registering and CRUD operations on ad-hoc data.