Add RawThetaSketchAggregationFunction (#5970)
Introduce `RawThetaSketchAggregationFunction` which collects the values for a given expression (can be single-valued or multi-valued) into a Sketch object, and returns the sketch as a base64 encoded string. It treats `BYTES` expression as serialized sketches.
It takes an optional second argument as the parameters for the function, e.g. `RAW_THETA_SKETCH(col, 'nominalEntries=8192')`.