[TE] Severity based alerter (#5962)
This PR adds the severity based alerting. It sends out anomalies to different email/Jira channels based on its anomaly severity. It also supports re-sending the alert if a given anomaly's severity has changed. This is achieved by adding a new lookup table for recording the relationship between an anomaly and subscription groups.