add `@types/react-router` dep (#6023)
Before, `@types/react-router` was unlisted. Because `@types/react-router-dom`
only has `@types/react-router: "*"` defined, it is possible the wrong version is
pulled. When this happens and a different major version is used, `npm run-script
build` would fail during maven builds over type errors.

After this, the correct version is now explicitly indicated to ensure the right
type version. For good measure, I hoisted the other `@types` dependencies to the
non-dev dependencies.

I'm not entirely sure what specific situations this happens in, but for my build
environment I was getting a different major version and maven builds were
failing. I suspect this is likely because of using an internal registry.