Enhance DistinctCountThetaSketchAggregationFunction (#6004)
## Description
Enhance `DistinctCountThetaSketchAggregationFunction`, and add the following supports:
- Support aggregating on raw values of all data types (both SV and MV)
- Support nested filter (E.g. `A = 1 AND (B = 2 OR C = 3)`)
- Support filter on all data types (both SV and MV)
- Support simple union without filters (E.g. `thetaSketch(col)`)
- Support `$0` as the default sketch (sketch without filter)

Also change the `DistinctCountRawThetaSketchAggregationFunction` to have the same usage as the `DistinctCountThetaSketchAggregationFunction` and returns base64 encoded sketch.
Fix the issue where theta-sketch parameters cannot be correctly handled in `BaseBrokerRequestHandler.updateColumnNames()`

## Release Notes
This aggregation function is still in beta version. This PR involves change on the format of data sent from server to broker, so it works only when both broker and server are upgraded to the new version.