Refresh the routing when realtime segment is committed (#6078)
Related to issue #6029 
When a consuming realtime segment is committed, it will refresh the segment ZK metadata (the partition metadata can change when the stream partition does not match the table config). The broker needs to refresh the routing for this segment in order to pick up the change in the segment ZK metadata.

This PR utilizes the `SegmentRefreshMessage` Helix message to refresh the broker routing when consuming realtime segment is committed.
Also add more tests in `SegmentPartitionLLCRealtimeClusterIntegrationTest` to verify the behavior.
With this change, when the stream partition changes for a partitioned table (which should be very rare), the result won't be accurate when the consuming segment contains records from multiple partitions, but will auto-recover when the segment is committed and the correct partition metadata is persisted to the segment ZK metadata.