Issue #4854 Framework for adding compatibility tests
This PR includes a basic framework for us to build compatibility tests across releases
The actual tests and a lot of code is yet to be written.

The framekwork is set up so that we can run operations during a cluster upgrade, and
detect incompatibility across releases. The value of the tests will be in the operations
we choose to test during upgrades

Configuration is done via YAML, and some sample yaml files and skinny classes have been
provided.