Added METADATA
METADATA file was added to prepare repository for enabling IMP.

The METADATA file was generated automatically by collecting top recent reviewers
of this repository. If reviewers did not exit - recent commit authors were added
as reviewers. It is possible this information is not accurate and it is
recommended to update the METADATA file accordingly.

See PHAB-272 for more details.
IMP docs: https://engdocs.uberinternal.com/IMP/
Support Slack room: #imp-on-boarding

Thank you!