Merge common APIs for Dictionary (#6176)
Motivation:
Currently the APIs for Dictionary is split in 3 places: `Dictionary`, `BaseImmutableDictionary`, `BaseMutableDictionary`. In order to use them, we need to cast the dictionary first, which is hard to manage and can potentially cause casting error.
E.g. #6174 is caused by casting an immutable dictionary to `BaseMutableDictionary`.
We should move the common read APIs to the root `Dictionary` interface to avoid the casting, and let all types of dictionary support these APIs.

Merge the following common APIs from `BaseImmutableDictionary` and `BaseMutableDictionary` to `Dictionary`:
- `insertionIndexOf`
- `getDictIdsInRange`
- `compare`
- `getMinVal`
- `getMaxVal`
- `getSortedValues`