Add StrictReplicaGroupInstanceSelector (#6208)
Recently we have introduced some features which require all the segments from the same partition to be processed by the same server. For example, upsert (#4261) feature requires all the segments for a partition loaded on a single server before starting serving queries to ensure result correctness. The Strict Replica-Group Routing is designed to meet this requirement.

This PR is the routing side change of the Strict Replica-Group Routing, which is handled in the StrictReplicaGroupInstanceSelector class.
The new algorithm relies on the ideal state of the table, so this PR added the ideal state to the routing interfaces.