[TE]frontend - Refactor detection-health and stats component (#6227)
Currently the detection-health component does not have the hosting container and tooltips. As a result it is individually configured in the calling component. Figured that part should remain the same regardless of which component instantiates the detection-health component. As a result, moving that static html inside the component.

The stats-cards component has duplicated code which could be removed by using the if template helper. We would be doing another round of refactoring here once the performance-api is ready.