[TE] fix changing createdTime of anomalies (#6269)
This PR is to fix a bug that the createdTime of anomalies is updated when new anomalies with earlier startTime is generated, which causes that the same anomalies get send out twice due to updated createdTime of anomalies. The fix is to merge new anomalies into existing anomalies if possible, regardless of its startTime.