[TE]frontend - Build new subroutes for single-metric-anomalies and composite-anomalies (#6263)
This PR adds support to navigate to the appropriate sub-routes for exploration of different anomaly types. Users would first be navigating to the /explore route when they select any alert from the home page or alert page. When the YAML is parsed in that route, we would have access to the type of the alert and thus we can subsequently redirect to the appropriate sub-route (/explore/single-metric-anomalies or /explore/composite-anomalies).

The changes below are also part of this PR:-

Added a util for computing the initial specification to pass into date-range-picker component. This is reusable so we'd be able to delete and greatly simplify the initialization code from all components/routes which instantiate date-range-picker as part of another PR.
Delete obsolete code
Refactored performance stats styling, using flexbox instead of display: inline-block. This helps reduce the css.
Styling improvement for the date-range-picker