[TE]frontend - Build the tree parser for composite anomalies (#6290)
The parser contains the logic to extract data in order to build child components at any level of the tree when performing the drill-down analysis for Entity Monitoring. The output at each level of drill-down contains data formatted in a way that can be directly consumed by the child components referenced in the output without any additional transformation.