Fix: Close HelixAdmin appropriately, when setting up Helix Cluster.
Currently, we cannot select multiple controller tests and run them in IDE successfully.
We see that after the first test that extends `ControllerTest`, the setup for next test
is unsuccessful due to the following reason:

- In `startController`, we setup Helix as well as Pinot Controllers.
- Both of these end up opening a ZKClient, however, only one of them (Pinot Controller)
  closes the zkClient.
- Since Helix controller does not close the ZKCLient, it remains alive, even after the
  first test's `cleanup()` is called.
- Given that we use SharedZKClientFactory which is a singleton class, and when running
  multiple tests from IDE, they use the same JVM, the second test ends up getting
  an obsolete ZKClient from preivous test, which is already disconnected.

This PR fixes the issue by appropriately closing the HelixAdmin (which closes the ZKClient)
when completed.