[TE]frontend - Add breadcrumb component (#6350)
As part of entity monitoring, we would need to display breadcrumbs as user drills down each level from the tree. This change adds the component for that.

Miscellaneous

Add composite-anomalies component which hosts the breadcrumb component and will host the parent anomalies component, group component for each level of drilldown.
Add support to disable date picker when user is not exploring anomalies at the root level.
Prevent eslint from being run on scss, css files. Eslint by default intented only for js files.
Prevent ember-template-lint from being run on hbs files. It is failing without output. I believe it is because we have not configured it yet. The package is not present in package.json.
Fix styling in couple of files.