Replace BrokerRequestOptimizer with QueryOptimizer to also optimize the PinotQuery (#6423)
Currently broker only optimizes the filter in BrokerRequest, but leaves the filter in PinotQuery as is. Because of this, we can only use the filter from the BrokerRequest, which is inefficient because there is no native Expression support. It also blocks the PQL deprecation.

In this PR:
- Attach the time boundary to PinotQuery.
- Replace the BrokerRequestOptimizer with QueryOptimizer to also optimize the filter in PinotQuery.
- in BrokerRequestToQueryContextConverter, use PinotQuery to generate QueryContext for SQL queries.

For backward-compatibility, Broker will optimize the filter in both BrokerRequest and PinotQuery.