[TE]frontend - Add filtering support for Entity Monitoring tables (#6514)
Filtering is supposed to work on column level with all possible data types in the table. It will do a substring match of the string passed in with each of the column entries; a return value of true would make table display the corresponding row.

Miscellaneous

Refactored anomaliesDetails to avoid doing the O(N) computation along with O(N) memory involved in converting object to array to be used in the template. Instead the template is modified to directly work with the object.
Refactored the columns in each of the table for entity monitoring to ensure reuse columns which can be shared.
Note to reviewers:
Lot of changes in utils.js are style related auto changed by the prettifier. Please focus on just 2 functions in that file -isObject and checkForMatch. Both are towards the end of the file.