[TE]frontend - Change the formatting of metric display in exploration page (#6545)
Composite alerts can have multiple metrics instead of just 1. The backend would be making the change to return a list of all metrics in yaml, instead of string. On the frontend we would be displaying the name of the first metric and the count of the remaining metrics for now.

If need be, we may show tooltips later to display the entire metric list, but at this point there is no decision made yet for that.