Optimize group-key generator (#6559)
Optimize the IntMapBasedHolder in the DictionaryBasedGroupKeyGenerator for better group-by performance on query with group-by columns cardinality product from 10K to 2B.
The improvement (up to ~40%) is mainly from replacing the Int2IntOpenHashMap with the new implemented IntGroupIdMap. The IntGroupIdMap stores both keys and values within a single array so that it is more friendly to CPU cache.