Directly store values in GroupKey and avoid ser/de (#6593)
Currently in the `GroupKeyGenerator.GroupKey` we serialized the keys into a string key, then deserialize the value back within `GroupByOrderByCombineOperator`. This PR modifies the GroupKey to directly store the keys without serializing them into a string to save the overhead of ser/de the group keys, and also reduce the garbage created.
Rename the existing `GroupKey` to `StringGroupKey` to keep the existing behavior of `GroupByCombineOperator` for PQL queries.