Remove the TableDataManager when there is no segment left for the table (#6608)
Current behavior: On server side, `TableDataManager` is created when loading the first segment for a table, but never removed during the lifetime of the server. This could cause problem when a table is deleted and recreated without restarting the server because server will treat the recreated table as the old table (e.g. upsert won't be enabled if it is not enabled in the old table but enabled in the new table).
The fix: When the last segment for a table is removed, also remove the `TableDataManager`.