Skip loading columns not in the schema (#6649)
Currently, when a column is added to the schema, server will generate default columns if the segment does not have the new column. But when a column is removed from the schema, the server still load all the columns within the segments. This could cause problem for `select *` queries because different segments might contain different columns.
With the change, server will skip loading the columns not in the schema, and all the segments should have the same set of columns loaded.