DataTable V3 implementation and measure data table serialization cost on server
DataTable V3 move metadata section to the end of bytes when
serialization and use enum values (instead of String in V2) as key.
This change will be backward incompat if servers are upgraded first,
so brokers must be upgraded before servers. The compatibility of
the protocols will not be retained beyond 0.8.0 (or the next
version that is released)