Unify data type conversion and formatting (#6728)
Unify the logic of data type conversion and formatting into `PinotDataType`, `DataType`, `ColumnDataType`.
- `PinotDataType` is used for type conversion
- `DataType` is used in schema to represent the source data type (only have single-valued type)
- `ColumnDataType` is used in `DataTable` and `ResultTable` to represent the result type

`BYTES` type is stored as `byte[]` externally but as `ByteArray` internally for hashing and sorting purpose. This PR handles the conversion of `BYTES` type (also new types to be added in the future) so that the the result can be passed to the UDF for post-aggregation calculations.
Also adding primitive array types to `PinotDataType`.

We should consider merging `DataType` and `ColumnDataType` in the future.