Interface changes for Kinesis connector (#6667)
Part 2 of Kinesis implementation changes. This #6518 is the main PR, which is being split up into 3 parts in order to reduce the scope of review.

In this PR, An abstraction has been introduced for a group of partition/shards, and each consumer will now be responsible for a PartitionGroup instead of just a partitionId (though in first iteration, the PartitionGroup will only contain 1 partition/shard). It includes corresponding additions to interfaces and some new interfaces.

Kafka stream should function as is, with no changes needed.

Note that before upgrading to this change, installations should upgrade to at least the release that has #5359 (https://docs.pinot.apache.org/basics/releases/0.5.0). An installation that has older versions than this may break during upgrade.